+++ crates/nostr/src/nips/nip49.rs
@@ -27,7 +27,6 @@ use crate::{key, SecretKey};
 const SALT_SIZE: usize = 16;
 const NONCE_SIZE: usize = 24;
 const CIPHERTEXT_SIZE: usize = 48;
-const TOTAL_SIZE: usize = 1 + 1 + SALT_SIZE + NONCE_SIZE + 1 + CIPHERTEXT_SIZE; // 91
 const KEY_SIZE: usize = 32;
 
 /// NIP49 error
@@ -188,6 +187,9 @@ pub struct EncryptedSecretKey {
 }
 
 impl EncryptedSecretKey {
+    /// Encrypted Secret Key len
+    pub const LEN: usize = 1 + 1 + SALT_SIZE + NONCE_SIZE + 1 + CIPHERTEXT_SIZE; // 91;
+
     /// Encrypt [SecretKey]
     #[inline]
     #[cfg(feature = "std")]
@@ -253,9 +255,9 @@ impl EncryptedSecretKey {
 
     /// Parse encrypted secret key from bytes
     pub fn from_slice(slice: &[u8]) -> Result<Self, Error> {
-        if slice.len() != TOTAL_SIZE {
+        if slice.len() != Self::LEN {
             return Err(Error::InvalidLength {
-                expected: TOTAL_SIZE,
+                expected: Self::LEN,
                 found: slice.len(),
             });
         }
@@ -302,7 +304,7 @@ impl EncryptedSecretKey {
 
     /// Get encrypted secret key as bytes
     pub fn as_vec(&self) -> Vec<u8> {
-        let mut bytes: Vec<u8> = Vec::with_capacity(TOTAL_SIZE);
+        let mut bytes: Vec<u8> = Vec::with_capacity(Self::LEN);
         bytes.push(self.version as u8);
         bytes.push(self.log_n);
         bytes.extend_from_slice(&self.salt);